% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_range.R
\name{draft_range}
\alias{draft_range}
\title{draft_range}
\usage{
draft_range(min, max)
}
\arguments{
\item{min}{First year in range.}

\item{max}{Last year in range.}
}
\value{
All draft picks from first year to last year in provided range.
}
\description{
This is general function that scrapes the pro-football-reference website for the year that you specify.
Will return all of the draft picks of that year with round, pick number, team, name, position, age, and college.
}
\examples{
draft_range(min = "2010", max = "2020")
}
