% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_year.R
\name{draft_year}
\alias{draft_year}
\title{draft_year}
\usage{
draft_year(year)
}
\arguments{
\item{year}{A year.}
}
\value{
All draft picks in provided year.
}
\description{
This is general function that scrapes the pro-football-reference website for the year that you specify.
Will return all of the draft picks of that year with round, pick number, team, name, position, age, and college.
}
\examples{
draft_year(year = "2020")

}
